git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@476 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/entitydef/entity_macro.hpp
@@ -129,7 +129,7 @@ public:																										\
 		{																									\
 			PropertyDescription* propertyDescription = iter->second;										\
 			PyObject* pyVal = PyDict_GetItemString(cellData, propertyDescription->getName().c_str());		\
-			(*mstream) << (uint32)propertyDescription->getUType();											\
+			(*mstream) << propertyDescription->getUType();													\
 			propertyDescription->getDataType()->addToStream(mstream, pyVal);								\
 		}																									\
 																											\
